Add new versions for LazyRow and FlowRow variants (#49)
This allows user to set a map of a key which value is a ColorPickerItem. These components can useful when we need to relate a custom object that needs to be selected with a ColorPickerItem.

Also, add documentation and apply a patch on `build.gradle.kts` to add `@samples` code block in sources.jar for multiplatform.